refactor(providers): extract shared OCR processing helpers
